Move SpaManager to Extensions namespace and enhance SPA handling
Relocated SpaManager from src/ to src/Extensions and updated ServiceProvider registration to use the new namespace. The new SpaManager adds authentication, permission, and CSRF checks, improves asset serving security, and streams large files efficiently. Removed obsolete .htaccess and the old SpaManager implementation.
