--- src/.htaccess
@@ -1,3 +1,0 @@
-Satisfy Any
-Order Allow,Deny
-Allow from all
